requirements.txt
python-dotenv
espn-api
pandas
tqdm
fastapi
uvicorn[standard]